frontend-utils: Extracted out structs from inside ExternalNavigatorBackend::fetch
